manual merge cli
