+++ lib/extIP.js
@@ -1,53 +1,103 @@
 'use strict';
-
-var request = require('request');
+var requests = require('./requests');
 var utils = require('./utils');
 
-module.exports.setup = function (config) {
+module.exports = function (extConf) {
+    extConf = extConf || {};
 
+    var isValid = utils.validateConfig(extConf);
 
-    var addValidation = function (request) {
-        return function (cb) {
-            return request(function (err, body) {
-                if (err) {
-                    return cb(err, null);
-                }
-                // if the body is null use an empty string
-                body = (body || '').toString().replace('\n', '');
-                return cb.apply(null, utils.isIP(body) ? [null, body] : [new Error('Invalid IP'), null]);
+    if (isValid.errors.length) {
+        console.error(isValid.errors);
+        process.exit(1);
+    }
 
-            });
-        };
+    // Check: https://github.com/mjhasbach/MOIRA
+    var defConf = {
+        getIP: 'sequential', // parallel
+        replace: false,
+        services: [
+            'http://ifconfig.co/x-real-ip',
+            'http://icanhazip.com/',
+            'http://ifconfig.me/ip',
+            'http://ip.appspot.com/',
+            'http://curlmyip.com/',
+            'http://ident.me/',
+            'http://whatismyip.akamai.com/',
+            'http://tnx.nl/ip',
+            'http://myip.dnsomatic.com/',
+            'http://ipecho.net/plain'
+        ],
+        timeout: 1000
     };
 
-    var requestFactory = function (request, url) {
-        return function (cb) {
-            return request.get({
-                url: url,
-                timeout: config.timeout,
-                headers: {
-                    'User-Agent': 'curl/'
-                }
-            }, function (err, res, body) {
-                cb.apply(null, err ? [err, null] : [null, body]);
-            });
-        };
-    };
+    var config = utils.mergeConfig(extConf, defConf);
 
-    var initializeServices = function (services) {
-        return services.map(function (url) {
-            return {
-                getIP: addValidation(requestFactory(request, url)),
-                url: url
+    var services = requests.setup(config).services;
+
+
+    var getIP = {
+        sequential: function (cb) {
+            var errors = [];
+            utils.asyncLoop({
+                iterations: services.length,
+                exec: function (i, stop, next) {
+                    services[i].getIP(function (err, ip) {
+                        if (err) {
+                            err = services[i].url + ' : ' + err;
+                            errors.push(err);
+                            next();
+                        } else {
+                            stop(ip);
+                        }
+                    });
+                },
+                done: function (ip) {
+                    cb.apply(null, ip ? [null, ip] : [errors, null]);
+                }
+            });
+        },
+
+        parallel: function (cb) {
+            var done = false;
+            var errors = [];
+            var requests;
+
+            var abort = function (requests) {
+                process.nextTick(function () {
+                    requests.forEach(function (request) {
+                        request.abort();
+                    });
+                });
             };
-        });
+
+
+            var onResponse = function (err, ip) {
+
+                if (done) {
+                    return;
+                }
+                if (err) {
+                    errors.push(err);
+                }
+                if(ip) {
+                    done = true;
+                    abort(requests); //async
+                    return cb(null, ip);
+                }
+                if (errors.length === services.length) {
+                    done = true;
+                    abort(requests); //async
+                    return cb(errors, null);
+                }
+            };
+
+            requests = services.map(function (service) {
+                return service.getIP(onResponse);
+            });
+
+        }
     };
 
-    return {
-        addValidation: addValidation,
-        requestFactory: requestFactory,
-        initializeServices: initializeServices,
-        services: initializeServices(config.services)
-    };
-
+    return getIP[config.getIP];
 };
